requirements.txt
UnityPy==1.10.18
fsb5
requests
PyGithub